+++ CHANGELOG.md
@@ -36,8 +36,14 @@
 ## 5.6.0
 Expected: March 2022
 
+In Clixon 5.6 the dependency of libevhtp for native RESTCONF is
+removed. Also module-state has been upgraded to RFC8525. A lot of bugs
+have been fixed, thanks to the community for all feedback.
+
 ### New features
 
+* Yang library upgraded from RFC78795 to RFC 8525
+  * See API changes for more info
 * HTTP/1 native parser as part of the RESTCONF client
   * Replaced libevhtp/libevent2 with internal http1 parser
   * Replace configure option `--disable-evhtp` with `--disable-http1` for disabling HTTP/1 whihc is on by default
@@ -46,10 +52,23 @@ Expected: March 2022
 
 Users may have to change how they access the system
 
+* Module state upgrade: RFC7895 to RFC 8525:
+  * To upgrade to RFC8525:
+    * You need to change `CLICON_YANG_LIBRARY` to `true` and `CLICON_MODULE_LIBRARY_RFC7895` to `false`
+  * To keep RFC7895:
+    * Set both `CLICON_YANG_LIBRARY` and `CLICON_MODULE_LIBRARY_RFC7895` to `true`
+  * This upgrade means that the state-data returned using GET is changed:
+    * preamble changed from: `<modules-state>...` to: `<yang-library><module-set>...`
+    * `module-state-id` changed to `content-id`
+    * `conformance-type` removed
+  * Note that the datastore feature `CLICON_XMLDB_MODSTATE` is backward compatible with RFC8525.
 * New `clixon-config@2022-02-11.yang` revision
   * Added option:
     * `CLICON_LOG_STRING_LIMIT`
-    * Removed (previosly marked) obsolete options:
+    * 'CLICON_YANG_LIBRARY`
+  * Changed default value:
+    * `CLICON_MODULE_LIBRARY_RFC7895` to false  
+  * Removed (previosly marked) obsolete options:
       * `CLICON_RESTCONF_PATH`
       * `CLICON_RESTCONF_PRETTY`
       * `CLICON_CLI_GENMODEL`
@@ -84,7 +103,7 @@ Users may have to change how they access the system
 ### Corrected Bugs
 
 * Fixed: [Validate error when appending module B grouping to module A item use augment statement #308](https://github.com/clicon/clixon/issues/308)
-* Fixed: [ Restconf PATCH method request failed on item defined by submodule #306](https://github.com/clicon/clixon/issues/306)
+* Fixed: [Restconf PATCH method request failed on item defined by submodule #306](https://github.com/clicon/clixon/issues/306)
 * Fixed: [restconf GET json response does not encode top level node with namespace as per rfc #303](https://github.com/clicon/clixon/issues/303)
 * Fixed: [statement: require-instance should be true if not present according to rfc7950 Sec 9.9.3](https://github.com/clicon/clixon/issues/302)
   * See also API changes